requirements.txt
fastapi==0.108.0
mysql_connector_repackaged==0.3.1
pydantic==2.5.3
python-dotenv==1.0.0
